# src/adapters/in_/http/routers/complete_order.py
# generated by fastapi-codegen:
#   filename:  openapi_contract.yaml

from __future__ import annotations

from uuid import UUID

from fastapi import APIRouter, Depends, Path
from that_depends import inject

from src.core.application.commands.complete_order import CompleteOrderCommand
from src.core.application.commands.complete_order import complete_order as complete_order_use_case
from src.core.domain.ports import UnitOfWork
from src.di.container import Container

from ..models import (
    Error,
)

router = APIRouter(tags=['CompleteOrder'])


@router.post(
    '/api/v1/couriers/{courierId}/orders/{orderId}/complete',
    response_model=None,
    status_code=200,
    responses={
        '400': {'model': Error},
        '409': {'model': Error},
        '500': {'model': Error},
    },
    tags=['CompleteOrder'],
)
@inject
async def complete_order(
    courier_id: UUID = Path(..., alias='courierId'),
    order_id: UUID = Path(..., alias='orderId'),
    uow: UnitOfWork = Depends(Container.unit_of_work),
) -> None:
    """
    Завершить заказ
    """
    cmd = CompleteOrderCommand(courier_id=courier_id, order_id=order_id)
    await complete_order_use_case(cmd, uow)