# src/adapters/in_/http/routers/create_courier.py
# generated by fastapi-codegen:
#   filename:  openapi_contract.yaml

from __future__ import annotations

from fastapi import APIRouter, Depends
from that_depends import inject

from src.core.application.commands.create_courier import CreateCourierCommand
from src.core.application.commands.create_courier import create_courier as create_courier_use_case
from src.core.domain.ports import UnitOfWork
from src.di.container import Container

from ..models import (
    CreateCourierResponse,
    Error,
    NewCourier,
)

router = APIRouter(tags=['CreateCourier'])


@router.post(
    '/api/v1/couriers',
    response_model=CreateCourierResponse,
    status_code=201,
    responses={
        '400': {'model': Error},
        '409': {'model': Error},
        '500': {'model': Error},
    },
    tags=['CreateCourier'],
)
@inject
async def create_courier(body: NewCourier, uow: UnitOfWork = Depends(Container.unit_of_work)) -> CreateCourierResponse:
    """
    Добавить курьера
    """
    cmd = CreateCourierCommand(name=body.name)
    courier = await create_courier_use_case(cmd, uow)

    return CreateCourierResponse(courierId=courier.id)