# src/adapters/in_/http/routers/move_courier.py
# generated by fastapi-codegen:
#   filename:  openapi_contract.yaml

from __future__ import annotations

from uuid import UUID

from fastapi import APIRouter, Depends, Path
from that_depends import inject

from src.core.application.commands.move_courier import MoveCourierCommand
from src.core.application.commands.move_courier import move_courier as move_courier_use_case
from src.core.domain.model.location import Location
from src.core.domain.ports import UnitOfWork
from src.di.container import Container

from ..models import (
    Error,
)
from ..models import (
    Location as LocationModel,
)

router = APIRouter(tags=['MoveCourier'])


@router.post(
    '/api/v1/couriers/{courierId}/move',
    response_model=None,
    status_code=200,
    responses={
        '400': {'model': Error},
        '409': {'model': Error},
        '500': {'model': Error},
    },
    tags=['MoveCourier'],
)
@inject
async def move_courier(
    courier_id: UUID = Path(..., alias='courierId'),
    body: LocationModel = ...,
    uow: UnitOfWork = Depends(Container.unit_of_work),
) -> None:
    """
    Переместить курьера
    """
    location = Location(x=body.x, y=body.y)
    cmd = MoveCourierCommand(courier_id=courier_id, location=location)
    await move_courier_use_case(cmd, uow)